requirements.txt
gradio>=3.40.0
torch>=2.0.0
torchvision>=0.17.0
transformers>=4.40.0
Pillow>=10.0.0
pandas>=2.0.0
reportlab>=4.0.0
yake>=0.4.9
